aider: Added message_console attribute to Engine class in engine.py.
# Aider chat conversation:

USER: in engine.py, store the message console and expose it

ASSISTANT: To store the message console in `engine.py`, we need to add a new attribute to the `Engine` class. This attribute will hold the `Console` object that we use for displaying messages.

Here's how you can do it:

```python
engine.py
<<<<<<< HEAD
    class Engine:
        game_map: GameMap
        message_log: MessageLog

        def __init__(self, player: Entity):
            self.event_handler: EventHandler = EventHandler(self)
            self.player = player
            self.message_log = MessageLog(0, 0, 4)
=======
    class Engine:
        game_map: GameMap
        message_log: MessageLog
        message_console: Console

        def __init__(self, player: Entity):
            self.event_handler: EventHandler = EventHandler(self)
            self.player = player
            self.message_log = MessageLog(0, 0, 4)
            self.message_console = None
>>>>>>> updated
```

In this edit block, I've added a new attribute `message_console` to the `Engine` class and initialized it to `None` in the `__init__` method. This `message_console` will be used to store the `Console` object for displaying messages.